requirements.txt
gunicorn
asgiref==3.7.2
Django==4.2.6
django-render-partial==0.4
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
psycopg-binary==3.1.12

djangorestframework~=3.14.0